+++ importer/importer/config/settings.py
@@ -1,9 +1,12 @@
 import os
 
-from services.fhir_keycloak_api import (FhirKeycloakApi, FhirKeycloakApiOptions, ExternalAuthenticationOptions,
-                                        InternalAuthenticationOptions)
-
 from dotenv import load_dotenv
+
+from importer.services.fhir_keycloak_api import (ExternalAuthenticationOptions,
+                                                 FhirKeycloakApi,
+                                                 FhirKeycloakApiOptions,
+                                                 InternalAuthenticationOptions)
+
 load_dotenv()
 
 client_id = os.getenv("client_id")
@@ -23,14 +26,26 @@ refresh_token = os.getenv("refresh_token")
 authentication_options = None
 if username is not None and password is not None:
     authentication_options = InternalAuthenticationOptions(
-        client_id=client_id, client_secret=client_secret, keycloak_base_uri=keycloak_url, realm=realm,
-        user_username=username, user_password=password)
+        client_id=client_id,
+        client_secret=client_secret,
+        keycloak_base_uri=keycloak_url,
+        realm=realm,
+        user_username=username,
+        user_password=password,
+    )
 elif access_token is not None and refresh_token is not None:
     authentication_options = ExternalAuthenticationOptions(
-        client_id=client_id, client_secret=client_secret, keycloak_base_uri=keycloak_url, realm=realm,
-        access_token=access_token, refresh_token=refresh_token)
+        client_id=client_id,
+        client_secret=client_secret,
+        keycloak_base_uri=keycloak_url,
+        realm=realm,
+        access_token=access_token,
+        refresh_token=refresh_token,
+    )
 else:
     raise ValueError("Unable to get authentication parameters")
 
-api_service_options = FhirKeycloakApiOptions(fhir_base_uri=fhir_base_url, authentication_options=authentication_options)
+api_service_options = FhirKeycloakApiOptions(
+    fhir_base_uri=fhir_base_url, authentication_options=authentication_options
+)
 api_service = FhirKeycloakApi(api_service_options)
