+++ importer/pytest.ini
@@ -1,0 +1,9 @@
+[pytest]
+env =
+    client_id = 'example-client-id'
+    client_secret = 'example-client-secret'
+    fhir_base_url = 'https://example.smartregister.org/fhir'
+    keycloak_url = 'https://keycloak.smartregister.org/auth'
+    realm = 'realm
+    username = 'example-username'
+    password = 'example-password'
