+++ importer/requirements.txt
@@ -4,9 +4,10 @@ requests==2.31.0
 requests-oauthlib==1.3.1
 urllib3==2.0.7
 backoff==2.2.1
-pytest==7.4.2
+pytest>=7.4.2
 jsonschema==4.21.1
 mock==5.1.0
 python-magic==0.4.27
 jwt
 python-dotenv==1.0.1
+pytest-env==1.1.3
