--- importer/sample_config.py
@@ -1,9 +1,0 @@
-client_id = 'example-client-id'
-client_secret = 'example-client-secret'
-username = 'example-username'
-password = 'example-password'
-access_token_url = 'https://keycloak.smartregister.org/auth/realms/example-realm/protocol/openid-connect/token'
-fhir_base_url = 'https://example.smartregister.org/fhir'
-keycloak_url = 'https://keycloak.smartregister.org/auth/admin/realms/example-realm'
-access_token = 'example-access-token'
-product_access_token = 'example-product-access-token'
