+++ importer/tests/test_builder.py
@@ -1,81 +1,28 @@
-import json
 import unittest
-from jsonschema import validate
+import json
+import pathlib
 from mock import patch
+from jsonschema import validate
 
-import os
-# TODO - Implement pytest-env or something cleaner than this for env vars
-os.environ["username"] = "username"
-os.environ["password"] = "password"
-os.environ["client_id"] = "client_id"
-os.environ["client_secret"] = "client_secret"
-os.environ["fhir_base_url"] = 'https://example.smartregister.org/fhir'
-os.environ["keycloak_url"] = 'https://keycloak.smartregister.org/auth'
-os.environ["product_access_token"] = 'example-product-access-token'
-os.environ["access_token"] = "access_token"
-os.environ["refresh_token"] = "refresh_token"
-os.environ["realm"] = "realm"
+from importer.builder import (build_payload, extract_resources, process_resources_list, extract_matches,
+                                   build_org_affiliation, check_parent_admin_level, build_assign_payload)
+from importer.utils import read_csv
 
-from main import (
-    read_csv,
-    write_csv,
-    build_payload,
-    build_org_affiliation,
-    extract_matches,
-    create_user_resources,
-    export_resources_to_csv,
-    build_assign_payload,
-    create_user,
-    confirm_keycloak_user,
-    confirm_practitioner,
-    check_parent_admin_level,
-    split_chunk,
-    read_file_in_chunks,
-    extract_resources,
-    process_resources_list,
-)
+dir_path = str(pathlib.Path(__file__).parent.resolve())
+csv_path = dir_path + "/../csv/"
+json_path = dir_path + "/../json_payloads/"
 
 
-class TestMain(unittest.TestCase):
-    def test_read_csv(self):
-        csv_file = "csv/users.csv"
-        records = read_csv(csv_file)
-        self.assertIsInstance(records, list)
-        self.assertEqual(len(records), 3)
-
-    def test_write_csv(self):
-        self.test_data = [
-            [
-                "e2e-mom",
-                "True",
-                "update",
-                "caffe509-ae56-4d42-945e-7b4c161723d1",
-                "d93ae7c3-73c0-43d1-9046-425a3466ecec",
-            ],
-            [
-                "e2e-skate",
-                "True",
-                "update",
-                "2d4feac9-9ab5-4585-9b33-e5abd14ceb0f",
-                "58605ed8-7217-4bf3-8122-229b6f47fa64",
-            ],
-        ]
-        self.test_resource_type = "test_organization"
-        self.test_fieldnames = ["name", "active", "method", "id", "identifier"]
-        csv_file = write_csv(
-            self.test_data, self.test_resource_type, self.test_fieldnames
-        )
-        csv_content = read_csv(csv_file)
-        self.assertEqual(csv_content, self.test_data)
-
-    @patch("main.get_resource")
+class TestBuilder(unittest.TestCase):
+    @patch("importer.builder.get_resource")
     def test_build_payload_organizations(self, mock_get_resource):
         mock_get_resource.return_value = "1"
 
-        csv_file = "csv/organizations/organizations_full.csv"
+        csv_file = csv_path + "organizations/organizations_full.csv"
         resource_list = read_csv(csv_file)
+        json_file = json_path + "organizations_payload.json"
         payload = build_payload(
-            "organizations", resource_list, "json_payloads/organizations_payload.json"
+            "organizations", resource_list, json_file
         )
         payload_obj = json.loads(payload)
         self.assertIsInstance(payload_obj, dict)
@@ -106,10 +53,10 @@ class TestMain(unittest.TestCase):
         validate(payload_obj["entry"][2]["request"], request_schema)
 
         #  TestCase organizations_min.csv
-        csv_file = "csv/organizations/organizations_min.csv"
+        csv_file = csv_path + "organizations/organizations_min.csv"
         resource_list = read_csv(csv_file)
         payload = build_payload(
-            "organizations", resource_list, "json_payloads/organizations_payload.json"
+            "organizations", resource_list, json_file
         )
         payload_obj = json.loads(payload)
         self.assertIsInstance(payload_obj, dict)
@@ -139,19 +86,19 @@ class TestMain(unittest.TestCase):
         }
         validate(payload_obj["entry"][0]["resource"], request_schema)
 
-    @patch("main.check_parent_admin_level")
-    @patch("main.get_resource")
+    @patch("importer.builder.check_parent_admin_level")
+    @patch("importer.builder.get_resource")
     def test_build_payload_locations(
         self, mock_get_resource, mock_check_parent_admin_level
     ):
         mock_get_resource.return_value = "1"
         mock_check_parent_admin_level.return_value = "3"
 
-        csv_file = "csv/locations/locations_full.csv"
+        csv_file = csv_path + "locations/locations_full.csv"
         resource_list = read_csv(csv_file)
+        json_file = json_path + "locations_payload.json"
         payload = build_payload(
-            "locations", resource_list, "json_payloads/locations_payload.json",
-            None, "http://terminology.hl7.org/CodeSystem/location-type"
+            "locations", resource_list, json_file, None, "http://terminology.hl7.org/CodeSystem/location-type"
         )
         payload_obj = json.loads(payload)
         self.assertIsInstance(payload_obj, dict)
@@ -237,11 +184,10 @@ class TestMain(unittest.TestCase):
         validate(payload_obj["entry"][0]["request"], request_schema)
 
         #  TestCase locations_min.csv
-        csv_file = "csv/locations/locations_min.csv"
+        csv_file = csv_path + "locations/locations_min.csv"
         resource_list = read_csv(csv_file)
         payload = build_payload(
-            "locations", resource_list, "json_payloads/locations_payload.json",
-            None, "http://terminology.hl7.org/CodeSystem/location-type"
+            "locations", resource_list, json_file, None, "http://terminology.hl7.org/CodeSystem/location-type"
         )
         payload_obj = json.loads(payload)
         self.assertIsInstance(payload_obj, dict)
@@ -271,8 +217,8 @@ class TestMain(unittest.TestCase):
         }
         validate(payload_obj["entry"][0]["resource"], request_schema)
 
-    @patch("main.handle_request")
-    @patch("main.get_base_url")
+    @patch("importer.builder.handle_request")
+    @patch("importer.builder.get_base_url")
     def test_check_parent_admin_level(self, mock_get_base_url, mock_handle_request):
         mock_get_base_url.return_value = "https://example.smartregister.org/fhir"
         mocked_response_text = {
@@ -297,18 +243,18 @@ class TestMain(unittest.TestCase):
         }
         string_mocked_response_text = json.dumps(mocked_response_text)
         mock_handle_request.return_value = (string_mocked_response_text, 200)
-        locationParentId = "18fcbc2e-4240-4a84-a270-7a444523d7b6"
-        admin_level = check_parent_admin_level(locationParentId)
+        location_parent_id = "18fcbc2e-4240-4a84-a270-7a444523d7b6"
+        admin_level = check_parent_admin_level(location_parent_id)
         self.assertEqual(admin_level, "3")
 
-    @patch("main.get_resource")
+    @patch("importer.builder.get_resource")
     def test_build_payload_care_teams(self, mock_get_resource):
         mock_get_resource.return_value = "1"
 
-        csv_file = "csv/careteams/careteam_full.csv"
+        csv_file = csv_path + "careteams/careteam_full.csv"
         resource_list = read_csv(csv_file)
         payload = build_payload(
-            "careTeams", resource_list, "json_payloads/careteams_payload.json"
+            "careTeams", resource_list, json_path + "careteams_payload.json"
         )
         payload_obj = json.loads(payload)
         self.assertIsInstance(payload_obj, dict)
@@ -398,16 +344,16 @@ class TestMain(unittest.TestCase):
         }
         validate(payload_obj["entry"][0]["request"], request_schema)
 
-    @patch("main.save_image")
-    @patch("main.get_resource")
+    @patch("importer.builder.save_image")
+    @patch("importer.builder.get_resource")
     def test_build_payload_group(self, mock_get_resource, mock_save_image):
         mock_get_resource.return_value = "1"
         mock_save_image.return_value = "f374a23a-3c6a-4167-9970-b10c16a91bbd"
 
-        csv_file = "csv/import/product.csv"
+        csv_file = csv_path + "import/product.csv"
         resource_list = read_csv(csv_file)
         payload, list_resource = build_payload(
-            "Group", resource_list, "json_payloads/product_group_payload.json", []
+            "Group", resource_list, json_path + "product_group_payload.json", []
         )
         payload_obj = json.loads(payload)
         self.assertEqual(list_resource, ['Binary/f374a23a-3c6a-4167-9970-b10c16a91bbd'])
@@ -468,7 +414,7 @@ class TestMain(unittest.TestCase):
 
         resource = [["Supply Inventory List", "current", "create", "77dae131-fd5d-4585-95db-2dd2b569d7a1"]]
         result_payload = build_payload(
-            "List", resource, "json_payloads/group_list_payload.json")
+            "List", resource, json_path + "product_list_payload.json")
         full_list_payload = process_resources_list(result_payload, created_resources)
 
         resource_schema = {
@@ -487,7 +433,7 @@ class TestMain(unittest.TestCase):
         validate(full_list_payload["entry"][0]["resource"], resource_schema)
 
     def test_extract_matches(self):
-        csv_file = "csv/organizations/organizations_locations.csv"
+        csv_file = csv_path + "organizations/organizations_locations.csv"
         resource_list = read_csv(csv_file)
         resources = extract_matches(resource_list)
         expected_matches = {
@@ -501,7 +447,7 @@ class TestMain(unittest.TestCase):
         self.assertEqual(resources, expected_matches)
 
     def test_build_org_affiliation(self):
-        csv_file = "csv/organizations/organizations_locations.csv"
+        csv_file = csv_path + "organizations/organizations_locations.csv"
         resource_list = read_csv(csv_file)
         resources = extract_matches(resource_list)
         payload = build_org_affiliation(resources, resource_list)
@@ -509,74 +455,6 @@ class TestMain(unittest.TestCase):
         self.assertIsInstance(payload_obj, dict)
         self.assertEqual(payload_obj["resourceType"], "Bundle")
         self.assertEqual(len(payload_obj["entry"]), 2)
-
-    def test_uuid_generated_in_creating_user_resources_is_unique_and_repeatable(self):
-        users = [
-            [
-                "Jane",
-                "Doe",
-                "Janey",
-                "jdoe@example.com",
-                "",
-                "Practitioner",
-                "true",
-                "a715b562-27f2-432a-b1ba-e57db35e0f93",
-                "test",
-                "demo",
-                "pa$$word",
-            ],
-            [
-                "John",
-                "Doe",
-                "Janey",
-                "jodoe@example.com",
-                "",
-                "Practitioner",
-                "true",
-                "a715b562-27f2-432a-b1ba-e57db35e0f93",
-                "test",
-                "demo",
-                "pa$$word",
-            ],
-            [
-                "Janice",
-                "Doe",
-                "Jenn",
-                "jendoe@example.com",
-                "99d54e3c-c26f-4500-a7f9-3f4cb788673f",
-                "Supervisor",
-                "false",
-                "a715b562-27f2-432a-b1ba-e57db35e0f93",
-                "test",
-                "demo",
-                "pa$$word",
-            ],
-        ]
-
-        users_uuids = {}
-        for user_id, user in enumerate(users):
-            payload = create_user_resources(user[4], user)
-            payload_obj = json.loads(payload)
-            practitioner_uuid = payload_obj["entry"][0]["resource"]["id"]
-            group_uuid = payload_obj["entry"][1]["resource"]["id"]
-            practitioner_role_uuid = payload_obj["entry"][2]["resource"]["id"]
-            users_uuids[user_id] = [
-                practitioner_uuid,
-                group_uuid,
-                practitioner_role_uuid,
-            ]
-
-        # practitioner_uuid
-        self.assertEqual(users_uuids[0][0], users_uuids[1][0])
-        self.assertNotEqual(users_uuids[1][0], users_uuids[2][0])
-
-        # group_uuid
-        self.assertEqual(users_uuids[0][1], users_uuids[1][1])
-        self.assertNotEqual(users_uuids[1][1], users_uuids[2][1])
-
-        # practitioner_role_uuid
-        self.assertEqual(users_uuids[0][2], users_uuids[1][2])
-        self.assertNotEqual(users_uuids[1][2], users_uuids[2][2])
 
     def test_uuid_generated_for_locations_is_unique_and_repeatable(self):
         resources = [
@@ -628,8 +506,8 @@ class TestMain(unittest.TestCase):
         ]
 
         payload = build_payload(
-            "locations", resources, "json_payloads/locations_payload.json",
-            None, "http://terminology.hl7.org/CodeSystem/location-type"
+            "locations", resources, json_path + "locations_payload.json", None,
+            "http://terminology.hl7.org/CodeSystem/location-type"
         )
         payload_obj = json.loads(payload)
         location1 = payload_obj["entry"][0]["resource"]["id"]
@@ -680,13 +558,13 @@ class TestMain(unittest.TestCase):
         ]
         with self.assertRaises(ValueError) as raised_error:
             build_payload(
-                "locations", resources, "json_payloads/locations_payload.json"
+                "locations", resources, json_path + "locations_payload.json"
             )
         self.assertEqual(
             "The id is required to update a resource", str(raised_error.exception)
         )
 
-    @patch("main.get_resource")
+    @patch("importer.builder.get_resource")
     def test_update_resource_with_non_existing_id_fails(self, mock_get_resource):
         mock_get_resource.return_value = "0"
         non_existing_id = "123"
@@ -706,81 +584,14 @@ class TestMain(unittest.TestCase):
         ]
         with self.assertRaises(ValueError) as raised_error:
             build_payload(
-                "locations", resources, "json_payloads/locations_payload.json"
+                "locations", resources, json_path + "locations_payload.json"
             )
         self.assertEqual(
             "Trying to update a Non-existent resource", str(raised_error.exception)
         )
 
-    @patch("main.write_csv")
-    @patch("main.handle_request")
-    @patch("main.get_base_url")
-    def test_export_resource_to_csv(
-        self, mock_get_base_url, mock_handle_request, mock_write_csv
-    ):
-        mock_get_base_url.return_value = "https://example.smartregister.org/fhir"
-        mock_response_data = {
-            "entry": [
-                {
-                    "resource": {
-                        "name": "City1",
-                        "status": "active",
-                        "id": "ba787982-b973-4bd5-854e-eacbe161e297",
-                        "identifier": [
-                            {"value": "ba787 982-b973-4bd5-854e-eacbe161e297"}
-                        ],
-                        "partOf": {
-                            "display": "test location-1",
-                            "reference": "Location/18fcbc2e-4240-4a84-a270"
-                            "-7a444523d7b6",
-                        },
-                        "type": [
-                            {"coding": [{"display": "Jurisdiction", "code": "jdn"}]}
-                        ],
-                        "physicalType": {
-                            "coding": [{"display": "Jurisdiction", "code": "jdn"}]
-                        },
-                    }
-                }
-            ]
-        }
-        string_response = json.dumps(mock_response_data)
-        mock_response = (string_response, 200)
-        mock_handle_request.return_value = mock_response
-        test_data = [
-            [
-                "City1",
-                "active",
-                "update",
-                "ba787982-b973-4bd5-854e-eacbe161e297",
-                "ba787 982-b973-4bd5-854e-eacbe161e297",
-                "test location-1",
-                "18fcbc2e-4240-4a84-a270-7a444523d7b6",
-                "Jurisdiction",
-                "jdn",
-                "Jurisdiction",
-                "jdn",
-            ]
-        ]
-        test_elements = [
-            "name",
-            "status",
-            "method",
-            "id",
-            "identifier",
-            "parentName",
-            "parentID",
-            "type",
-            "typeCode",
-            "physicalType",
-            "physicalTypeCode",
-        ]
-        resource_type = "Location"
-        export_resources_to_csv("Location", "_lastUpdated", "gt2023-08-01", 1)
-        mock_write_csv.assert_called_once_with(test_data, resource_type, test_elements)
-
-    @patch("main.handle_request")
-    @patch("main.get_base_url")
+    @patch("importer.builder.handle_request")
+    @patch("importer.builder.get_base_url")
     def test_build_assign_payload_update_assigned_org(
         self, mock_get_base_url, mock_handle_request
     ):
@@ -843,8 +654,8 @@ class TestMain(unittest.TestCase):
             payload_obj["entry"][0]["resource"]["organization"]["display"], "New Org"
         )
 
-    @patch("main.handle_request")
-    @patch("main.get_base_url")
+    @patch("importer.builder.handle_request")
+    @patch("importer.builder.get_base_url")
     def test_build_assign_payload_create_org_assignment(
         self, mock_get_base_url, mock_handle_request
     ):
@@ -899,8 +710,8 @@ class TestMain(unittest.TestCase):
             payload_obj["entry"][0]["resource"]["organization"]["display"], "New Org"
         )
 
-    @patch("main.handle_request")
-    @patch("main.get_base_url")
+    @patch("importer.builder.handle_request")
+    @patch("importer.builder.get_base_url")
     def test_build_assign_payload_create_new_practitioner_role(
         self, mock_get_base_url, mock_handle_request
     ):
@@ -942,8 +753,8 @@ class TestMain(unittest.TestCase):
             payload_obj["entry"][0]["resource"]["organization"]["display"], "New Org"
         )
 
-    @patch("main.handle_request")
-    @patch("main.get_base_url")
+    @patch("importer.builder.handle_request")
+    @patch("importer.builder.get_base_url")
     def test_build_assign_payload_create_new_link_location_to_inventory_list(
         self, mock_get_base_url, mock_handle_request
     ):
@@ -984,8 +795,8 @@ class TestMain(unittest.TestCase):
             "Location/3af23539-850a-44ed-8fb1-d4999e2145ff",
         )
 
-    @patch("main.handle_request")
-    @patch("main.get_base_url")
+    @patch("importer.builder.handle_request")
+    @patch("importer.builder.get_base_url")
     def test_build_assign_payload_update_location_with_new_inventory(
         self, mock_get_base_url, mock_handle_request
     ):
@@ -1042,8 +853,8 @@ class TestMain(unittest.TestCase):
             "Group/e62a049f-8d48-456c-a387-f52e72c39c74",
         )
 
-    @patch("main.handle_request")
-    @patch("main.get_base_url")
+    @patch("importer.builder.handle_request")
+    @patch("importer.builder.get_base_url")
     def test_build_assign_payload_create_new_link_location_to_inventory_list_with_multiples(
         self, mock_get_base_url, mock_handle_request
     ):
@@ -1100,460 +911,3 @@ class TestMain(unittest.TestCase):
             payload_obj["entry"][1]["resource"]["entry"][0]["item"]["reference"],
             "Group/c0666a5a-00f6-488c-9001-8630560b5810",
         )
-
-    @patch("main.logging")
-    @patch("main.handle_request")
-    @patch("main.get_keycloak_url")
-    def test_create_user(
-        self, mock_get_keycloak_url, mock_handle_request, mock_logging
-    ):
-        mock_get_keycloak_url.return_value = (
-            "https://keycloak.smartregister.org/auth/admin/realms/example-realm"
-        )
-        mock_handle_request.return_value.status_code = 201
-        mock_handle_request.return_value.headers = {
-            "Location": "https://keycloak.smartregister.org/auth/admin/realms"
-            "/example-realm/users/6cd50351-3ddb-4296-b1db"
-            "-aac2273e35f3"
-        }
-        mocked_user_data = (
-            "Jenn",
-            "Doe",
-            "Jenny",
-            "jeendoe@example.com",
-            "431cb523-253f-4c44-9ded-af42c55c0bbb",
-            "Supervisor",
-            "TRUE",
-            "a715b562-27f2-432a-b1ba-e57db35e0f93",
-            "test",
-            "demo",
-            "pa$$word",
-        )
-        user_id = create_user(mocked_user_data)
-
-        self.assertEqual(user_id, "6cd50351-3ddb-4296-b1db-aac2273e35f3")
-        mock_logging.info.assert_called_with("Setting user password")
-
-    @patch("main.handle_request")
-    @patch("main.get_keycloak_url")
-    def test_create_user_already_exists(
-        self, mock_get_keycloak_url, mock_handle_request
-    ):
-        mock_get_keycloak_url.return_value = (
-            "https://keycloak.smartregister.org/auth/admin/realms/example-realm"
-        )
-        mock_handle_request.return_value.status_code = 409
-        mocked_user_data = (
-            "Jenn",
-            "Doe",
-            "Jenn",
-            "jendoe@example.com",
-            " 99d54e3c-c26f-4500-a7f9-3f4cb788673f",
-            "Supervisor",
-            "false",
-            "a715b562-27f2-432a-b1ba-e57db35e0f93",
-            "test",
-            "demo",
-            "pa$$word",
-        )
-        user_id = create_user(mocked_user_data)
-        self.assertEqual(user_id, 0)
-
-    # Test the confirm_keycloak function
-    @patch("main.logging")
-    @patch("main.handle_request")
-    @patch("main.get_keycloak_url")
-    def test_confirm_keycloak_user(
-        self, mock_get_keycloak_url, mock_handle_request, mock_logging
-    ):
-        mock_get_keycloak_url.return_value = (
-            "https://keycloak.smartregister.org/auth/admin/realms/example-realm"
-        )
-        mocked_user_data = (
-            "Jenn",
-            "Doe",
-            "Jenny",
-            "jeendoe@example.com",
-            "431cb523-253f-4c44-9ded-af42c55c0bbb",
-            "Supervisor",
-            "TRUE",
-            "a715b562-27f2-432a-b1ba-e57db35e0f93",
-            "test",
-            "demo",
-            "pa$$word",
-        )
-        user_id = create_user(mocked_user_data)
-        self.assertEqual(user_id, 0)
-
-        mock_response = (
-            '[{"id":"6cd50351-3ddb-4296-b1db-aac2273e35f3","createdTimestamp":1710151827166,'
-            '"username":"Jenny","enabled":true,"totp":false,"emailVerified":false,"firstName":"Jenn",'
-            '"lastName":"Doe","email":"jeendoe@example.com","attributes":{"fhir_core_app_id":["demo"]},'
-            '"disableableCredentialTypes":[],"requiredActions":[],"notBefore":0,"access":{'
-            '"manageGroupMembership":true,"view":true,"mapRoles":true,"impersonate":true,'
-            '"manage":true}}]',
-            200,
-        )
-        mock_handle_request.return_value = mock_response
-        mock_json_response = json.loads(mock_response[0])
-        keycloak_id = confirm_keycloak_user(mocked_user_data)
-
-        self.assertEqual(mock_json_response[0]["username"], "Jenny")
-        self.assertEqual(mock_json_response[0]["email"], "jeendoe@example.com")
-        mock_logging.info.assert_called_with("User confirmed with id: " + keycloak_id)
-
-    # Test confirm_practitioner function
-    @patch("main.handle_request")
-    @patch("main.get_base_url")
-    def test_confirm_practitioner_if_practitioner_uuid_not_provided(
-        self, mock_get_base_url, mock_handle_request
-    ):
-        mock_get_base_url.return_value = "https://example.smartregister.org/fhir"
-        mocked_user = (
-            "Jenn",
-            "Doe",
-            "Jenny",
-            "jeendoe@example.com",
-            "",
-            "Supervisor",
-            "TRUE",
-            "a715b562-27f2-432a-b1ba-e57db35e0f93",
-            "test",
-            "demo",
-            "pa$$word",
-        )
-        mocked_response_data = {
-            "resourceType": "Bundle",
-            "type": "searchset",
-            "total": 1,
-        }
-        string_response = json.dumps(mocked_response_data)
-        mock_response = (string_response, 200)
-        mock_handle_request.return_value = mock_response
-        practitioner_exists = confirm_practitioner(
-            mocked_user, "431cb523-253f-4c44-9ded-af42c55c0bbb"
-        )
-        self.assertTrue(
-            practitioner_exists, "Practitioner exist, linked to the provided user"
-        )
-
-    @patch("main.logging")
-    @patch("main.handle_request")
-    @patch("main.get_base_url")
-    def test_confirm_practitioner_linked_keycloak_user_and_practitioner(
-        self, mock_get_base_url, mock_handle_request, mock_logging
-    ):
-        mock_get_base_url.return_value = "https://example.smartregister.org/fhir"
-        mocked_user = (
-            "Jenn",
-            "Doe",
-            "Jenny",
-            "jeendoe@example.com",
-            "6cd50351-3ddb-4296-b1db-aac2273e35f3",
-            "Supervisor",
-            "TRUE",
-            "a715b562-27f2-432a-b1ba-e57db35e0f93",
-            "test",
-            "demo",
-            "pa$$word",
-        )
-        mocked_response_data = {
-            "resourceType": "Practitioner",
-            "identifier": [
-                {"use": "official", "value": "431cb523-253f-4c44-9ded-af42c55c0bbb"},
-                {"use": "secondary", "value": "6cd50351-3ddb-4296-b1db-aac2273e35f3"},
-            ],
-        }
-        string_response = json.dumps(mocked_response_data)
-        mock_response = (string_response, 200)
-        mock_handle_request.return_value = mock_response
-        practitioner_exists = confirm_practitioner(
-            mocked_user, "6cd50351-3ddb-4296-b1db-aac2273e35f3"
-        )
-        self.assertTrue(practitioner_exists)
-        self.assertEqual(
-            mocked_response_data["identifier"][1]["value"],
-            "6cd50351-3ddb-4296-b1db-aac2273e35f3",
-        )
-        mock_logging.info.assert_called_with(
-            "The Keycloak user and Practitioner are linked as expected"
-        )
-
-    # Test create_user_resources function
-    def test_create_user_resources(self):
-        user = (
-            "Jenn",
-            "Doe",
-            "Jenn",
-            "jendoe@example.com",
-            "99d54e3c-c26f-4500-a7f9-3f4cb788673f",
-            "Supervisor",
-            "false",
-            "a715b562-27f2-432a-b1ba-e57db35e0f93",
-            "test",
-            "demo",
-            "pa$$word",
-        )
-        user_id = "99d54e3c-c26f-4500-a7f9-3f4cb788673f"
-        payload = create_user_resources(user_id, user)
-        payload_obj = json.loads(payload)
-        self.assertIsInstance(payload_obj, dict)
-        self.assertEqual(payload_obj["resourceType"], "Bundle")
-        self.assertEqual(len(payload_obj["entry"]), 3)
-
-        resource_schema = {
-            "type": "object",
-            "properties": {
-                "resourceType": {"const": "Practitioner"},
-                "id": {"const": "99d54e3c-c26f-4500-a7f9-3f4cb788673f"},
-                "identifier": {
-                    "type": "array",
-                    "items": {
-                        "type": "object",
-                        "properties": {
-                            "use": {
-                                "type": "string",
-                                "enum": ["official", "secondary"],
-                            },
-                            "type": {
-                                "type": "object",
-                                "properties": {
-                                    "coding": {
-                                        "type": "array",
-                                        "items": {
-                                            "type": "object",
-                                            "properties": {
-                                                "system": {
-                                                    "const": "http://hl7.org/fhir/identifier-type"
-                                                },
-                                                "code": {"const": "KUID"},
-                                                "display": {
-                                                    "const": "Keycloak user ID"
-                                                },
-                                            },
-                                        },
-                                    },
-                                    "text": {"const": "Keycloak user ID"},
-                                },
-                            },
-                            "value": {"const": "99d54e3c-c26f-4500-a7f9-3f4cb788673f"},
-                        },
-                    },
-                },
-                "name": {
-                    "type": "array",
-                    "items": {
-                        "type": "object",
-                        "properties": {
-                            "use": {"const": "official"},
-                            "family": {"const": "Doe"},
-                            "given": {"type": "array", "items": {"type": "string"}},
-                        },
-                    },
-                },
-            },
-            "required": ["resourceType", "id", "identifier", "name"],
-        }
-        validate(payload_obj["entry"][0]["resource"], resource_schema)
-
-        request_schema = {
-            "type": "object",
-            "properties": {
-                "method": {"const": "PUT"},
-                "url": {"const": "Practitioner/99d54e3c-c26f-4500-a7f9-3f4cb788673f"},
-                "ifMatch": {"const": "1"},
-            },
-        }
-        validate(payload_obj["entry"][0]["request"], request_schema)
-
-        resource_schema = {
-            "type": "object",
-            "properties": {
-                "resourceType": {"const": "Group"},
-                "id": {"const": "0de5f541-65ca-5504-ad6b-9b386e5f8810"},
-                "identifier": {"type": "array", "items": {"type": "object"}},
-                "name": {"const": "Jenn Doe"},
-                "member": {
-                    "type": "array",
-                    "items": {
-                        "type": "object",
-                        "properties": {
-                            "entity": {
-                                "type": "object",
-                                "properties": {
-                                    "reference": {
-                                        "const": "Practitioner/99d54e3c-c26f-4500-a7f9-3f4cb788673f"
-                                    }
-                                },
-                            }
-                        },
-                    },
-                },
-            },
-            "required": ["resourceType", "id", "identifier", "name", "member"],
-        }
-        validate(payload_obj["entry"][1]["resource"], resource_schema)
-
-        request_schema = {
-            "type": "object",
-            "properties": {
-                "method": {"const": "PUT"},
-                "url": {"const": "Group/0de5f541-65ca-5504-ad6b-9b386e5f8810"},
-                "ifMatch": {"const": "1"},
-            },
-        }
-        validate(payload_obj["entry"][1]["request"], request_schema)
-
-        resource_schema = {
-            "type": "object",
-            "properties": {
-                "resourceType": {"const": "PractitionerRole"},
-                "id": {"const": "f08e0373-932e-5bcb-bdf2-0c28a3c8fdd3"},
-                "identifier": {"type": "array", "items": {"type": "object"}},
-                "practitioner": {
-                    "type": "object",
-                    "properties": {
-                        "reference": {
-                            "const": "Practitioner/99d54e3c-c26f-4500-a7f9-3f4cb788673f"
-                        },
-                        "display": {"const": "Jenn Doe"},
-                    },
-                },
-                "code": {
-                    "type": "object",
-                    "properties": {
-                        "coding": {
-                            "type": "array",
-                            "items": {
-                                "type": "object",
-                                "properties": {
-                                    "system": {"const": "http://snomed.info/sct"},
-                                    "code": {"const": "236321002"},
-                                    "display": {"const": "Supervisor (occupation)"},
-                                },
-                            },
-                        }
-                    },
-                },
-            },
-            "required": ["resourceType", "id", "identifier", "practitioner", "code"],
-        }
-        validate(payload_obj["entry"][2]["resource"], resource_schema)
-
-        request_schema = {
-            "type": "object",
-            "properties": {
-                "method": {"const": "PUT"},
-                "url": {
-                    "const": "PractitionerRole/f08e0373-932e-5bcb-bdf2-0c28a3c8fdd3"
-                },
-                "ifMatch": {"const": "1"},
-            },
-        }
-        validate(payload_obj["entry"][2]["request"], request_schema)
-
-    @patch("main.set_resource_list")
-    def test_split_chunk_direct_sync_first_chunk_less_than_size(
-        self, mock_set_resource_list
-    ):
-        chunk = '[{"id": "10", "resourceType": "Patient"}'
-        next_left_over = split_chunk(chunk, "", 50, {}, "direct")
-        chunk_list = '[{"id": "10", "resourceType": "Patient"}]'
-        self.assertEqual(next_left_over, "-")
-        mock_set_resource_list.assert_called_once_with(chunk_list)
-
-    @patch("main.set_resource_list")
-    def test_split_chunk_direct_sync_middle_chunk_less_than_size(
-        self, mock_set_resource_list
-    ):
-        chunk = ' "resourceType": "Patient"}'
-        left_over_chunk = '{"id": "10",'
-        next_left_over = split_chunk(chunk, left_over_chunk, 50, {}, "direct")
-        chunk_list = '[{"id": "10", "resourceType": "Patient"}]'
-        self.assertEqual(next_left_over, "-")
-        mock_set_resource_list.assert_called_once_with(chunk_list)
-
-    @patch("main.set_resource_list")
-    def test_split_chunk_direct_sync_last_chunk_less_than_size(
-        self, mock_set_resource_list
-    ):
-        left_over_chunk = '{"id": "10", "resourceType": "Patient"}]'
-        next_left_over = split_chunk("", left_over_chunk, 50, {}, "direct")
-        chunk_list = '[{"id": "10", "resourceType": "Patient"}]'
-        self.assertEqual(next_left_over, "-")
-        mock_set_resource_list.assert_called_once_with(chunk_list)
-
-    @patch("main.set_resource_list")
-    def test_split_chunk_direct_sync_first_chunk_greater_than_size(
-        self, mock_set_resource_list
-    ):
-        chunk = '[{"id": "10", "resourceType": "Patient"},{"id": "11", "resourceType":'
-        next_left_over = split_chunk(chunk, "", 40, {}, "direct")
-        chunk_list = '[{"id": "10", "resourceType": "Patient"}]'
-        self.assertEqual(next_left_over, '{"id": "11", "resourceType":')
-        mock_set_resource_list.assert_called_once_with(chunk_list)
-
-    @patch("main.set_resource_list")
-    def test_split_chunk_direct_sync_middle_chunk_greater_than_size(
-        self, mock_set_resource_list
-    ):
-        chunk = ': "Task"},{"id": "10", "resourceType": "Patient"},{"id": "11", "resourceType":'
-        left_over_chunk = '{"id": "09", "resourceType"'
-        next_left_over = split_chunk(chunk, left_over_chunk, 80, {}, "direct")
-        chunk_list = '[{"id": "09", "resourceType": "Task"},{"id": "10", "resourceType": "Patient"}]'
-        self.assertEqual(next_left_over, '{"id": "11", "resourceType":')
-        mock_set_resource_list.assert_called_once_with(chunk_list)
-
-    @patch("main.set_resource_list")
-    def test_split_chunk_direct_sync_last_chunk_greater_than_size(
-        self, mock_set_resource_list
-    ):
-        left_over_chunk = '{"id": "10", "resourceType": "Patient"},{"id": "11", "resourceType": "Task"}]'
-        next_left_over = split_chunk("", left_over_chunk, 43, {}, "direct")
-        chunk_list = '[{"id": "10", "resourceType": "Patient"},{"id": "11", "resourceType": "Task"}]'
-        self.assertEqual(next_left_over, "")
-        mock_set_resource_list.assert_called_once_with(chunk_list)
-
-    @patch("main.set_resource_list")
-    @patch("main.build_resource_type_map")
-    def test_split_chunk_sort_sync_first_chunk_less_than_size(
-        self, mock_build_resource_type_map, mock_set_resource_list
-    ):
-        chunk = '[{"id": "10", "resourceType": "Patient"},{"id": "11"'
-        next_left_over = split_chunk(chunk, "", 50, {}, "sort")
-        chunk_list = '[{"id": "10", "resourceType": "Patient"}]'
-        self.assertEqual(next_left_over, '{"id": "11"')
-        mock_set_resource_list.assert_not_called()
-        mock_build_resource_type_map.assert_called_once_with(chunk_list, {}, 0)
-
-    def test_build_resource_type_map(self):
-        json_file = "tests/json/sample.json"
-        mapping = read_file_in_chunks(json_file, 300, "sort")
-        mapped_resources = {
-            "Patient": [0],
-            "Practitioner": [1, 5],
-            "Location": [2, 4],
-            "Observation": [3],
-        }
-        self.assertIsInstance(mapping, dict)
-        self.assertEqual(mapping, mapped_resources)
-
-    @patch("main.check_parent_admin_level")
-    @patch("main.get_resource")
-    def test_define_own_location_type_coding_system_url(self, mock_get_resource, mock_check_parent_admin_level):
-        mock_get_resource.return_value = "1"
-        mock_check_parent_admin_level.return_value = "3"
-        test_system_code = "http://terminology.hl7.org/CodeSystem/test_location-type"
-
-        csv_file = "csv/locations/locations_full.csv"
-        resource_list = read_csv(csv_file)
-        payload = build_payload(
-            "locations", resource_list, "json_payloads/locations_payload.json",
-            None, test_system_code
-        )
-        payload_obj = json.loads(payload)
-        self.assertEqual(payload_obj["entry"][0]["resource"]["type"][0]["coding"][0]["system"], test_system_code)
-
-
-if __name__ == "__main__":
-    unittest.main()
